Fix User Balance Bug
Fixed a lot of stuff to this point, but haven't got to the chart yet xD
